pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "fastapi>=0.123.9",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "user-agents>=2.2.0",
    "openpyxl>=3.1.5",
    "mnemonic>=0.21",
    "httpx>=0.28.1",
    "africastalking>=2.0.1",
    "python-dotenv>=1.2.1",
]
